Extrapolation.py: account for random h5py group order
For some files with some versions of h5py, the list of groups from
an h5py.File() is in a (seemingly) random order. We sort the output
of this list so that it is at least in a consistent order,
regardless of h5py's crazy behavior.

In particular, this issue was seen with Matt's 0.994 rerun (Lev5)
with h5py from orca's python 2.7.{6,8} modules.